fix(FE, BE): 새 파일을 업로드하지 않을 시 zod 경고문이 뜨는 문제 해결
- 문제 해결을 위해 formSchema를 다시 ReviewForm 컴포넌트 내로 옮김
- BE 단에서 deletedImages가 없을 경우에 에러 뜨는 소소한 이슈 해결

+++ back/controllers/Review.js
@@ -117,11 +117,13 @@ export const updateReview = asyncHandler(async (req, res) => {
   }
 
   // 새 이미지 파일이 있을 경우 기존 이미지 파일들 삭제 후 경로 업데이트
-  updateData.deletedImages.forEach((imagePath) => fs.unlinkSync(imagePath));
+  if (updateData.deletedImages) {
+    updateData.deletedImages.forEach((imagePath) => fs.unlinkSync(imagePath));
 
-  reviewData.images = reviewData.images.filter(
-    (image) => !updateData.deletedImages.includes(image)
-  );
+    reviewData.images = reviewData.images.filter(
+      (image) => !updateData.deletedImages.includes(image)
+    );
+  }
 
   if (req.files && req.files.length > 0) {
     reviewData.images = reviewData.images.concat(
